pyproject.toml
[project]
name = "audai"
version = "0.1.0"
description = "AI Audio CLI - TTS, Whisper, and Audio-to-Audio conversation tool"
authors = [
    { name = "Chophe", email = "chophe@yahoo.com" }
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "keyboard>=0.13.5"
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/audai"]

[project.scripts]
audai = "audai.cli:main_cli"
